requirements.txt
fastapi==0.99.1
pydantic>=2.9.2
tortoise-orm
aerich
uvicorn==0.20.0